+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+Add basic crafts with `Refreshable`, `AutoRefreshing`, and `Persistable` functionality, and
+a combination of these.
+
 ## 0.0.0
 
-- Pub.dev package reservation
+Pub.dev package reservation